chore: move ResponsiveNavLink to UI library
