+++ resources/js/Components/ui/index.tsx
@@ -34,6 +34,7 @@ export * from './Popover'
 export * from './Progress'
 export * from './RadioGroup'
 export * from './Resizable'
+export * from './ResponsiveNavLink'
 export * from './ScrollArea'
 export * from './Select'
 export * from './Separator'
